requirements.txt
openai
langchain
qdrant-client
tiktoken
python-dotenv
streamlit
